Rename network/utils/block.ts to network/utils/serializers.ts
