+++ ironfish/src/network/messages/newBlockV2.ts
@@ -4,7 +4,7 @@
 import bufio from 'bufio'
 import { SerializedCompactBlock } from '../../primitives/block'
 import { NetworkMessageType } from '../types'
-import { getCompactBlockSize, readCompactBlock, writeCompactBlock } from '../utils/block'
+import { getCompactBlockSize, readCompactBlock, writeCompactBlock } from '../utils/serializers'
 import { NetworkMessage } from './networkMessage'
 
 export class NewBlockV2Message extends NetworkMessage {
